% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bozukyol.R
\name{get_bozukyol}
\alias{get_bozukyol}
\title{get_bozukyol Function}
\usage{
get_bozukyol(busdoornumber)
}
\arguments{
\item{busdoornumber}{This parameter is not mandatory. This information is the information that comes to the bus driver screens. For this reason, you can also query bad roads by the bus door number.}
}
\description{
You can reach instant bad road information. See \url{https://data.ibb.gov.tr/organization/iett-genel-mudurlugu} for more information.
}
\examples{
 get_bozukyol(busdoornumber="B4042")

}
