% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_durakdetay.R
\name{get_durakdetay}
\alias{get_durakdetay}
\title{get_durakdetay Function}
\usage{
get_durakdetay(linecode)
}
\arguments{
\item{linecode}{You can make a query with/without bus line code.}
}
\description{
You can reach line code, line name, line situation, area and journey time instantly with this function. See \url{https://data.ibb.gov.tr/organization/iett-genel-mudurlugu} for more information.
}
\examples{
 get_durakdetay(linecode="145T")

}
