% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_duyuru.R
\name{get_duyuru}
\alias{get_duyuru}
\title{get_duyuru Function}
\usage{
get_duyuru()
}
\description{
You can reach daily announcements by type and bus line. See \url{https://data.ibb.gov.tr/organization/iett-genel-mudurlugu} for more information.
}
\examples{
 get_duyuru()

}
