% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hatdetay.R
\name{get_hatdetay}
\alias{get_hatdetay}
\title{get_hatdetay Function}
\usage{
get_hatdetay(linecode)
}
\arguments{
\item{linecode}{You can make a query with/without bus line code.}
}
\description{
You can reach all the route and bus stop information instantly with this function. See \url{https://data.ibb.gov.tr/organization/iett-genel-mudurlugu} for more information.
}
\examples{
 get_hatdetay(linecode="145T")

}
