% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ispark.R
\name{get_ispark}
\alias{get_ispark}
\title{get_ispark Function}
\usage{
get_ispark()
}
\description{
You can reach location and speed (km/h) of any bus with/without number plate. See \url{https://data.ibb.gov.tr/organization/iett-genel-mudurlugu} for more information.
}
