% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lokasyon.R
\name{get_lokasyon}
\alias{get_lokasyon}
\title{get_lokasyon Function}
\usage{
get_lokasyon(numberplate)
}
\arguments{
\item{numberplate}{This is not mandatory. You can query with only number plate.}
}
\description{
You can reach location and speed (km/h) of any bus with/without number plate. See \url{https://data.ibb.gov.tr/organization/iett-genel-mudurlugu} for more information.
}
\examples{
 get_lokasyon(numberplate="34 NL 7572")

}
