% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_guzergah.R, R/plot_lokasyon.R
\name{\%>\%}
\alias{\%>\%}
\alias{plot_guzergah}
\alias{plot_lokasyon}
\title{plot_guzergah Function}
\usage{
plot_guzergah(linecode)

plot_lokasyon(numberplate)
}
\arguments{
\item{linecode}{This is mandatory. You can query bus line with bus line code.}

\item{numberplate}{This is mandatory. You can query bus with bus' number plate.}
}
\description{
You can see the bus line route that you want in a map. You can reach all bus lines with get_hatdetay() function.    See \url{https://data.ibb.gov.tr/organization/iett-genel-mudurlugu} for more information.

You can see the bus location that you want in a map. See \url{https://data.ibb.gov.tr/organization/iett-genel-mudurlugu} for more information.
}
\examples{
 plot_guzergah(linecode="145T")

 plot_lokasyon(numberplate="34 NL 7572")

}
