% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_guzergah.R
\name{plot_guzergah}
\alias{plot_guzergah}
\title{plot_guzergah Function}
\usage{
plot_guzergah(linecode)
}
\arguments{
\item{linecode}{This is mandatory. You can query bus line with bus line code.}
}
\description{
You can see the bus line route that you want in a map. See \url{https://data.ibb.gov.tr/organization/iett-genel-mudurlugu} for more information.
}
\examples{
 plot_guzergah(linecode="145T")

}
