% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lokasyon.R
\name{plot_lokasyon}
\alias{plot_lokasyon}
\title{plot_lokasyon Function}
\usage{
plot_lokasyon(numberplate)
}
\arguments{
\item{numberplate}{This is mandatory. You can query bus with bus' number plate.}
}
\description{
You can see the bus location that you want in a map. See \url{https://data.ibb.gov.tr/organization/iett-genel-mudurlugu} for more information.
}
\examples{
 plot_lokasyon(numberplate="34 NL 7572")

}
